requirements.txt
streamlit>=1.29.0
google-generativeai>=0.8.0